split unit and integration tests for blueprint
